use mergeWithReplaceArrays to merge array values on add-on yaml editor
