+++ shell/edit/provisioning.cattle.io.cluster/rke2.vue
@@ -22,11 +22,10 @@ import { _CREATE, _EDIT, _VIEW } from '@shell/config/query-params';
 import { findBy, removeObject, clear } from '@shell/utils/array';
 import { createYaml } from '@shell/utils/create-yaml';
 import {
-  clone, diff, set, get, isEmpty
+  clone, diff, set, get, isEmpty, mergeWithReplaceArrays
 } from '@shell/utils/object';
 import { allHash } from '@shell/utils/promise';
 import { sortBy } from '@shell/utils/sort';
-import { vspherePoolConfigMerge } from '@shell/machine-config/vmwarevsphere-pool-config-merge';
 
 import { compare, sortable } from '@shell/utils/version';
 import { isHarvesterSatisfiesVersion, labelForAddon } from '@shell/utils/cluster';
@@ -1260,7 +1259,7 @@ export default {
         delete clonedCurrentConfig.metadata;
 
         if (this.provider === VMWARE_VSPHERE) {
-          machinePool.config = vspherePoolConfigMerge(clonedLatestConfig, clonedCurrentConfig);
+          machinePool.config = mergeWithReplaceArrays(clonedLatestConfig, clonedCurrentConfig);
         } else {
           machinePool.config = merge(clonedLatestConfig, clonedCurrentConfig);
         }
@@ -1644,7 +1643,7 @@ export default {
       const defaultChartValue = this.versionInfo[name];
       const key = this.chartVersionKey(name);
 
-      return merge({}, defaultChartValue?.values || {}, this.userChartValues[key] || {});
+      return mergeWithReplaceArrays(defaultChartValue?.values, this.userChartValues[key]);
     },
 
     initServerAgentArgs() {
